requirements.txt
streamlit
PyPDF2
langchain
langchain-openai
langchain-huggingface
langchain-community
sentence-transformers
faiss-cpu
python-dotenv